requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-crispy-forms==1.9.0
django-filter==2.2.0
django-formtools==2.2
django-guardian==2.2.0
django-import-export==2.0.2
django-reversion==3.0.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.9
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.3
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
sqlparse==0.3.1
tablib==1.1.0
uritemplate==3.0.1
urllib3==1.25.8
xadmin2==2.0.3
xlrd==1.2.0
xlwt==1.3.0
